+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- target bucket stage is obtained from `JANIS_ENV` environment variable
+### Changed
+- target bucket now is a janis-trace-service bucket
+- `add` method now receives client code name instead bucket name
+
 ## [1.1.0] - 2019-10-08
 ### Added
 - log structure validations